feat: update imports to use unified utils structure and rename size types
