+++ src/components/badge/Badge.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import {Badge} from "./Badge";
 import {Button} from "../button/Button";
 import {IconGitBranch} from "@tabler/icons-react";
-import {Colors as BadgeVariants} from "../../utils/component";
+import {Colors as BadgeVariants} from "../../utils";
 
 const meta: Meta = {
     title: "Badge",
