+++ src/components/quote/Quote.stories.tsx
@@ -1,7 +1,7 @@
 import {Meta, StoryObj} from "@storybook/react-vite";
 import {Quote} from "./Quote";
 import React from "react";
-import {Colors} from "../../utils/component";
+import {Colors} from "../../utils";
 
 const meta: Meta = {
     title: "Quote",
