+++ src/components/spacing/Spacing.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import {Component, Code0Sizes, mergeComponentProps} from "../../utils";
+import {Component, Sizes, mergeComponentProps} from "../../utils";
 import "./Spacing.style.scss"
 
 export interface SpacingProps extends Component<HTMLDivElement> {
-    spacing: Code0Sizes
+    spacing: Sizes
 }
 
 export const Spacing: React.FC<SpacingProps> = (props) => {
